requirements.txt
# Core ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Model Training and Evaluation
xgboost>=2.0.0
lightgbm>=4.1.0
imbalanced-learn>=0.11.0

# Fairness and Bias Detection
fairlearn>=0.9.0
aif360>=0.5.0

# Drift Detection
alibi-detect>=0.11.0
evidently>=0.4.0

# API and Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Dashboard and Visualization
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Data Processing
openpyxl>=3.1.0
python-dateutil>=2.8.2

# Monitoring and Logging
mlflow>=2.8.0
loguru>=0.7.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Database (optional for production)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.6.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
requests>=2.31.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipykernel>=6.26.0
ipywidgets>=8.1.0
notebook>=7.0.0

